修改了自瞄中的姿态解算算法 移除了OpenCL支持

+++ include/Utils.hpp
@@ -4,6 +4,6 @@ namespace Math
 
 double PointDistance(const cv::Point2f& pt1, const cv::Point2f& pt2);
 
-void RegularizeErrAngle(float& angle, char type);
+void RegularizeErrAngle(double& angle, char type);
 
 } //Math
